refactor: make BaseServable to unify behaviors of Apps and serve functions
